Move Identity and GenericIdentity to special.py
