+++ sympy/matrices/expressions/__init__.py
@@ -6,8 +6,7 @@ from .companion import CompanionMatrix
 from .funcmatrix import FunctionMatrix
 from .inverse import Inverse
 from .matadd import MatAdd
-from .matexpr import (Identity, MatrixExpr, MatrixSymbol, OneMatrix,
-                      matrix_symbols)
+from .matexpr import MatrixExpr, MatrixSymbol, OneMatrix, matrix_symbols
 from .matmul import MatMul
 from .matpow import MatPow
 from .trace import Trace, trace
@@ -20,7 +19,7 @@ from .dotproduct import DotProduct
 from .kronecker import kronecker_product, KroneckerProduct, combine_kronecker
 from .permutation import PermutationMatrix, MatrixPermute
 from .sets import MatrixSet
-from .special import ZeroMatrix
+from .special import ZeroMatrix, Identity
 
 __all__ = [
     'MatrixSlice',
