+++ sympy/matrices/expressions/matexpr.py
@@ -846,105 +846,6 @@ class MatrixSymbol(MatrixExpr):
             )]
 
 
-class Identity(MatrixExpr):
-    """The Matrix Identity I - multiplicative identity
-
-    Examples
-    ========
-
-    >>> from sympy.matrices import Identity, MatrixSymbol
-    >>> A = MatrixSymbol('A', 3, 5)
-    >>> I = Identity(3)
-    >>> I*A
-    A
-    """
-
-    is_Identity = True
-
-    def __new__(cls, n):
-        n = _sympify(n)
-        cls._check_dim(n)
-
-        return super().__new__(cls, n)
-
-    @property
-    def rows(self):
-        return self.args[0]
-
-    @property
-    def cols(self):
-        return self.args[0]
-
-    @property
-    def shape(self):
-        return (self.args[0], self.args[0])
-
-    @property
-    def is_square(self):
-        return True
-
-    def _eval_transpose(self):
-        return self
-
-    def _eval_trace(self):
-        return self.rows
-
-    def _eval_inverse(self):
-        return self
-
-    def conjugate(self):
-        return self
-
-    def _entry(self, i, j, **kwargs):
-        eq = Eq(i, j)
-        if eq is S.true:
-            return S.One
-        elif eq is S.false:
-            return S.Zero
-        return KroneckerDelta(i, j, (0, self.cols-1))
-
-    def _eval_determinant(self):
-        return S.One
-
-    def _eval_power(self, exp):
-        return self
-
-
-class GenericIdentity(Identity):
-    """
-    An identity matrix without a specified shape
-
-    This exists primarily so MatMul() with no arguments can return something
-    meaningful.
-    """
-    def __new__(cls):
-        # super(Identity, cls) instead of super(GenericIdentity, cls) because
-        # Identity.__new__ doesn't have the same signature
-        return super(Identity, cls).__new__(cls)
-
-    @property
-    def rows(self):
-        raise TypeError("GenericIdentity does not have a specified shape")
-
-    @property
-    def cols(self):
-        raise TypeError("GenericIdentity does not have a specified shape")
-
-    @property
-    def shape(self):
-        raise TypeError("GenericIdentity does not have a specified shape")
-
-    # Avoid Matrix.__eq__ which might call .shape
-    def __eq__(self, other):
-        return isinstance(other, GenericIdentity)
-
-    def __ne__(self, other):
-        return not (self == other)
-
-    def __hash__(self):
-        return super().__hash__()
-
-
 class OneMatrix(MatrixExpr):
     """
     Matrix whose all entries are ones.
@@ -1188,4 +1089,4 @@ from .matadd import MatAdd
 from .matpow import MatPow
 from .transpose import Transpose
 from .inverse import Inverse
-from .special import ZeroMatrix
+from .special import ZeroMatrix, Identity
