+++ sympy/matrices/expressions/matmul.py
@@ -7,12 +7,11 @@ from sympy.matrices.common import ShapeError, NonInvertibleMatrixError
 from sympy.matrices.matrices import MatrixBase
 
 from .inverse import Inverse
-from .matexpr import \
-    MatrixExpr, Identity, OneMatrix, GenericIdentity
+from .matexpr import MatrixExpr, OneMatrix
 from .matpow import MatPow
 from .transpose import transpose
 from .permutation import PermutationMatrix
-from .special import ZeroMatrix
+from .special import ZeroMatrix, Identity, GenericIdentity
 
 
 # XXX: MatMul should perhaps not subclass directly from Mul
